Fix erroneous decimals in price formatter.

+++ reko/reko/formatters.py
@@ -5,7 +5,7 @@ def format_price(price: Decimal) -> str:
     kronor = int(price)
     ore = price % 1
 
-    ore_str = str(ore * 100).zfill(2) if ore else "-"
+    ore_str = str(int(ore * 100)).zfill(2) if ore else "-"
 
     return f"{kronor}:{ore_str}"
